moves confidence logic to HelperFunctions adjusts modulenavigator
